add full 64bit support
so far, our 64 bit support was only half-baked in that we could
correctly read archives with 64 bit indices, however internal
bookkeeping was all done using unsigned integers (32 bit), and
so we relied on that no file/game ever would become > 4GB.
as we can see with StrangeLand, even the main exe already ships
with 3 GB, plus another GB of audio in separate files, so it's
not unlikely game authors will continue the unfortunate trend
to produce ever more bloated games, replacing quality with
quantity.

tested with a 4 GB empty file injected as first file of Strangeland,
so all game-related files were past 4GB, with packing, unpacking,
disassembling, assembling, injecting.

+++ Clib32.h
@@ -21,8 +21,8 @@ struct MultiFileLibNew {
 	char data_filenames[MAXMULTIFILES][50];
 	size_t num_data_files;
 	char filenames[MAX_FILES][100];
-	unsigned offset[MAX_FILES];
-	unsigned length[MAX_FILES];
+	unsigned long long offset[MAX_FILES];
+	unsigned long long length[MAX_FILES];
 	char file_datafile[MAX_FILES];        // number of datafile
 	size_t num_files;
 };
